reorganized sidebar sub somponents
