+++ src/styles/blog/ArticlePage.scss
@@ -2,7 +2,4 @@
   .article {
     padding-top: 25px;
   }
-  .email-subscribe {
-    text-align: right;
-  }
 }
