+++ src/styles/blog/BlogPage.scss
@@ -1,9 +1,3 @@
 .blog-page {
-  a {
-    color: black;
-    &:hover, &:focus {
-      text-decoration: none;
-    }
-  }
 
 }
