+++ src/styles/blog/sections/SideBar.scss
@@ -1,3 +1,5 @@
 .component-side-bar {
-
+   a{
+     color: black;
+   }
 }
